Allow odd stride values
They are actually quite useful for testing unaligned behavior: does it
slow down every access, or does it slow down only some accesses?

For example, on my i7-6700K, I get

	[torvalds@i7 test-tlb]$ ./test-tlb 16k 12
	  1.29ns (~5.0 cycles)

	[torvalds@i7 test-tlb]$ ./test-tlb 16k 11
	  1.36ns (~5.3 cycles)

which means that the unaligned load case is actually very low in pointer
chasing situations.  With a stride of 11, 75% of all the memory accesses
will be unaligned, but it only costs an additional 0.3 cycles on average.

The reason is that unaligned loads are basically free, except when they
cross some other boundary (*).  To test what the cost of that is, I'd
need to add a "offset" parameter and allow people to force every access
to be a bank crosser.

Some day.

(*) There may be other costs, like just resource use costs.  But the
kinds of boundaries that tend to matter are L1 D$ boundaries (that need
to access two lines), but also obviously page-crossers (that need two
TLB entries), and also sub-banks in the L1 (depending on width of the L1
interface).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>