Improve on map re-allocation strategy
Don't bother reallocating hugepage mappings, the cache associativity
issues are hopefully not all that noticeable with big pages to begin
with, and with fewer big pages to go around it's likely less effective
to try to change the mapping anyway.

For the regular page case, force any remapping to over-map the old
mapping, which will free the old pages.  That might cause the new
mapping to have similar allocation patterns as the old one, but not
using excessive amounts of memory still makes this the better option.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>