Improve on the timing logic a bit more
Time the warmup run to make a better estimate of how long to run the
real runs.  We want to balance getting timely results with still
actually having the time to actually traverse the full list a few times.

This basically aims for the real run to traverse the list at least five
times, but to also run for at least a fifth of a second.  Because five
is obviously a holy random number, and is also how many times we run the
test to try to de-noise the final number.

This also changes "make run" to actuall test the 256M case like the
README says (rather than stop at 128M).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>